Apply engine/API review fixes + partial-repaint tear fix (52 findings)
Reactive core
- Changed/Invalidated handlers and Observe callbacks run untracked —
  handler reads no longer become phantom dependencies of the recomputing
  node (Reactive.UntrackedInvoke).
- Reaction source tracking rewritten to the ordered-array positional
  reconcile (Reactively algorithm); drops two eager HashSets per node.

Partial repaint correctness (user-visible hover/scroll artifact)
- New PaintSnapshot: on partial frames the re-walked paint list is diffed
  against what was last submitted; commands that changed without a damage
  mark widen the damage, unboundable changes force a full repaint. Fixes
  torn repaints from unmarked visual changes (tooltip position, per-walk
  text like the fps counter). RepaintTracker: AddDamageBoundsOnly /
  ForceFullDamage. Diagnostics behind ZIGOTE_DEBUG_DAMAGE=1.

UI runtime
- Tooltip no longer forces full-damage frames forever while hovered;
  drag ghost repositions via paint-time translate + scoped damage instead
  of full relayout + MarkAll per pointer move.
- Stack: tight-constraint fast path (was double-measuring every child and
  permanently defeating descendant measure caches).
- Label wraps by per-word measurement (was polluting the global text
  memo with every line prefix); wrap/layout cache keys include all style
  inputs. Container applies c.Constrain. ListView exposes only visible
  rows to focus/semantics. Navigator skips laying out covered routes.
  SelectableText hit-tests per-line instead of scanning all chars.

Core
- SetClipboard uses bounded stackalloc (unbounded payload could overflow
  the stack). UTF-8 text cache allocates on the pinned object heap — no
  per-frame GCHandle pinning. ModifierKeys gains Cmd=8 (matches native).

API (breaking where noted)
- Row/Column/Flex: spacing parameter. Renames with [Obsolete] forwarders:
  MainAxisAlignment/CrossAxisAlignment, Checkbox.Value, OnSubmitted,
  TabBar onChanged, SelectedIndex. SizedBox.Shrink is a factory method
  (was a shared mutable instance). WidgetState.OwnEffect ties Effect
  lifetime to widget disposal. TextEditingController.Changed is wired to
  the bound TextField. Silent no-op params implemented (Text.softWrap,
  Material ListView padding) or deleted (keyboardType,
  debugShowCheckedModeBanner) [breaking]. Slider/Stepper are float
  end-to-end [breaking]. Null onPressed uniformly renders disabled, live.

ECS / game runtime
- EcsWorld caches native queries per component set (ForEach no longer
  creates/destroys a flecs query per call); T3/T4 overloads; system
  exceptions logged once per slot (was silently swallowed) with lock-free
  dispatch; iterator leak on throwing callbacks fixed.
- Physics sync batched (one FFI call for all bodies, quaternions — no
  Euler round-trip). GameApp runs a real fixed-timestep accumulator with
  OnFixedStep [breaking: Component.OnFixedUpdate removed — OnUpdate IS
  the 120 Hz tick]. Time.InterpolationAlpha exposed. Script dispatch no
  longer allocates a closure per component per tick. SpatialHash stores
  positions in cells; rotation/scale changes invalidate it.

Infra
- Exported games no longer ship the DevTools overlay; opt in via
  DevToolsEnabled in the project manifest [breaking]. Dead project refs
  dropped; DevTools + Ecs.Benchmark added to the solution. Native build
  now watches *.m and vendored C sources (stale-dylib fix).
  Zigote.Network docs state its experimental, unwired status.
  InspectorPanel split into 6 partial files (pure code motion).

Tests: ~85 new (reactive untracking/reconcile, layout caches, paint diff,
Material API, ECS cache, DevTools export). Suite: 1039 passed; renderer
changes verified byte-identical via golden captures.

+++ Zigote.UI.Gallery/Pages/LocalizationPage.cs
@@ -28,7 +28,7 @@ internal sealed class LocalizationPage : StatefulWidget
 internal sealed class LocalizationPageState : WidgetState<LocalizationPage>
 {
     private static readonly string[] Genders = ["male", "female", "other"];
-    private double _count = 2;
+    private float _count = 2;
     private int _gender;
 
     public override Widget Build(BuildContext context)
